refactor: extract Kanban graph persistence into topical sibling
