+++ tests/hermes_cli/test_kanban_worktree_isolation.py
@@ -22,6 +22,7 @@ from pathlib import Path
 import pytest
 
 from hermes_cli import kanban_db as kb
+from hermes_cli.kanban_db_graph import decompose_triage_task
 from hermes_cli import kanban_db_connect as kbc
 from hermes_cli import kanban_db_workspace as kbw
 
@@ -78,7 +79,7 @@ def test_decompose_worktree_children_get_own_workspace(kanban_home):
         )
         conn.commit()
 
-        child_ids = kb.decompose_triage_task(
+        child_ids = decompose_triage_task(
             conn,
             root,
             root_assignee="orchestrator",
